+++ example/js/app.js
@@ -1,9 +1,9 @@
-var wrapper = document.getElementById("signature-pad"),
-    clearButton = wrapper.querySelector("[data-action=clear]"),
-    savePNGButton = wrapper.querySelector("[data-action=save-png]"),
-    saveSVGButton = wrapper.querySelector("[data-action=save-svg]"),
-    canvas = wrapper.querySelector("canvas"),
-    signaturePad;
+var wrapper = document.getElementById("signature-pad");
+var clearButton = wrapper.querySelector("[data-action=clear]");
+var savePNGButton = wrapper.querySelector("[data-action=save-png]");
+var saveSVGButton = wrapper.querySelector("[data-action=save-svg]");
+var canvas = wrapper.querySelector("canvas");
+var signaturePad = new SignaturePad(canvas);
 
 // Adjust canvas coordinate space taking into account pixel ratio,
 // to make it look crisp on mobile devices.
@@ -13,15 +13,22 @@ function resizeCanvas() {
     // some browsers report devicePixelRatio as less than 1
     // and only part of the canvas is cleared then.
     var ratio =  Math.max(window.devicePixelRatio || 1, 1);
+
+    // This part causes the canvas to be cleared
     canvas.width = canvas.offsetWidth * ratio;
     canvas.height = canvas.offsetHeight * ratio;
     canvas.getContext("2d").scale(ratio, ratio);
+
+    // This library does not listen for canvas changes, so after the canvas is automatically
+    // cleared by the browser, SignaturePad#isEmpty might still return false, even though the
+    // canvas looks empty, because the internal data of this library wasn't cleared. To make sure
+    // that the state of this library is consistent with visual state of the canvas, you
+    // have to clear it manually.
+    signaturePad.clear();
 }
 
 window.onresize = resizeCanvas;
 resizeCanvas();
-
-signaturePad = new SignaturePad(canvas);
 
 clearButton.addEventListener("click", function (event) {
     signaturePad.clear();